+++ leptos-fluent-macros/src/lib.rs
@@ -31,7 +31,6 @@ use languages::build_languages_quote;
 pub(crate) use languages::ParsedLanguage;
 use loader::{I18nLoader, Identifier, LitBoolExpr};
 use quote::quote;
-use std::str::FromStr;
 
 #[cfg(feature = "debug")]
 #[inline(always)]
@@ -103,28 +102,6 @@ pub(crate) fn debug(msg: &str) {
 pub fn leptos_fluent(
     input: proc_macro::TokenStream,
 ) -> proc_macro::TokenStream {
-    // optional double braces (must be removed in v0.2)
-    let input_as_str = input.to_string();
-    let loader_input;
-    let double_braced_warning;
-    if input_as_str.starts_with('{') && input_as_str.ends_with('}') {
-        loader_input = proc_macro::TokenStream::from(
-            proc_macro2::TokenStream::from_str(
-                &input_as_str[1..input_as_str.len() - 1],
-            )
-            .unwrap(),
-        );
-        let warning = proc_macro_warning::FormattedWarning::new_deprecated(
-            "leptos_fluent",
-            "The double braced syntax `leptos_fluent! {{ ... }}` is deprecated and will be removed in v0.2. Use `leptos_fluent! { ... }` instead.",
-            proc_macro2::Span::call_site(),
-        );
-        double_braced_warning = quote!(#warning);
-    } else {
-        loader_input = input;
-        double_braced_warning = quote!();
-    }
-
     let I18nLoader {
         fluent_file_paths,
         children,
@@ -180,7 +157,7 @@ pub fn leptos_fluent(
         initial_language_from_data_file,
         #[cfg(feature = "system")]
         data_file_key,
-    } = syn::parse_macro_input!(loader_input as I18nLoader);
+    } = syn::parse_macro_input!(input as I18nLoader);
 
     let n_languages = languages.len();
     let languages_quote = build_languages_quote(&languages);
@@ -276,7 +253,7 @@ pub fn leptos_fluent(
         };
 
         let effect_quote = quote! {
-            if let Ok(l) = ::leptos_fluent::current_locale() {
+            if let Ok(l) = ::leptos_fluent::current_locale::current_locale() {
                 lang = ::leptos_fluent::l(&l, &LANGUAGES);
                 if let Some(l) = lang {
                     #initial_language_from_system_to_data_file_quote
@@ -1933,7 +1910,6 @@ pub fn leptos_fluent(
         .collect();
 
     let quote = quote! {
-        #double_braced_warning
         let i18n = {
             const LANGUAGES: [&::leptos_fluent::Language; #n_languages] =
                 #languages_quote;