+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # CHANGELOG
 
-## [Unreleased]
+## [Unreleased] - 0.2.0
+
+### Breaking changes
+
+- The feature `json` is not enabled by default anymore. Now leptos-fluent
+  does not includes features by default, so you don't need to use
+  `default-features = false` in your `Cargo.toml` file.
 
 ## ? - [0.2.0]
 
@@ -383,7 +389,7 @@ leptos_fluent! {
   global attribute of the `<html>` tag with the current language direction.
 - Multiple translations can be passed to the `leptos_fluent!` macro.
 
-## Versioning
+### Versioning
 
 **leptos-fluent** will include breaking changes in minor versions during
 the v0.x.0 series until v1.0.0 version is reached. Is safe to pin the